Remove commented code, add some return values
Now that %httpd.reb has its own issue tracker, commented out bits of
code can be moved into issues.  This does that for a few remarks.

It also annotates some functions with return values to make them
clearer, and has some other tweaks that are in line with the "Ren-C
coding guidelines" that are sort of forming together.