Eliminate manual chunking, add stress test
For reasons that were not entirely understood, there was a workaround
needed in the httpd in order to perform manual chunking in the
WRITE operation.

The problem appears to be addressed by fixes in Ren-C.  So plain
WRITE is used instead.

To test that it's working, a stress test which creates larger and
larger strings is used--doubling them each time.