Add stress test that hangs up arbitrarily on client requests
Asynchronous PORT! operations previously did not have any way to trap
errors, because:

    trap [
        write async-port data
    ] then [
        ; the write happens later in the event/WAIT loop...
        ; ...once this trap is over, and has no influence
    ]

The beginnings of a generalized 'error EVENT! are now added to the
interpreter--enough that it's possible to have a server that should be
able to stay running even if the client disconnects at any point.
This changes to use that event, and adds a "hangup" option to the
stress test...which for about half of its requests will pick a random
(deterministic) moment to drop the connection.