Improve previous commit using /PART revocation
The previous commit wanted to copy the entire `headers/Content-Type`
in the case there was no semicolon in it.  To ensure that outcome, it
would mutate the field to add a semicolon at the end.  This may not
be desirable, and is unnecessary...since if FIND is unable to locate
the searched-for-character it will return NULL.  When COPY gets a
NULL for the /PART argument, it "revokes" the refinement, as if it
had not been used with /PART--hence copying the whole string.