Change to use tuples (dots) instead of slashes for fields
With Rebol languages already having it fairly mixed-up as to what is
a function call and what is a variable reference, it can be disorienting
with so many slashes to know what represents a refinement to a function
call and what isn't.

With TUPLE! as an alternate sequence type there is now a way to write
things with more variety, that enforces what's on the left of the dot
is not a function call.