Avert out of memory error when randomness fails
The random element in the stress test was in conflict with the test
termination condition, which was checking the total transferred bytes
at the end of the loop (that might not be reached an arbitrary number
of times based on the random hangups).

This led some random situations to go up to the point where it would
run out of memory.  That just hadn't been seen yet.

Move the check to the beginning of the loop.