Switch to NULL member initializations instead of _
The quoted NULL (a simple ') is an easy way to make a NULL with a
similar low profile to that of BLANK!.  This helps avoid conflating
empty things with the surrounding filled values.

However, NULL is a more "aggressive" form of intent than BLANK!, as
it will error if you try to use it with something like APPEND...while
BLANK! will signal a desire to opt out of many operations silently.

An even more aggressive intent can be achieved with `~` which will
mean variables can be written to, but not (easily) read.

Includes some more fixes of the / vs . usage on field selection.