Try to be robust to common READ hangups
If a request is being processed and the user does something in a browser
like refresh before it is finished, this would cause an
error in READ that was not being trapped.

Also seen was a case where a connection would be opened and then
closed--something like a ping, with no actual HTTP
request.  (It's not clear what caused this, but if the browser
disable cache was clicked it seemed to stop)

We want to know what's going wrong with the server and raise an alarm,
but as these are common when accessing with browsers
and that is a common expectation, this just
turns them into log messages for now.

They are recognized by strings which is not ideal.