Use TRY on LENGTH inside ATTEMPT
ATTEMPT has been modified to be able to return NULL now, and the TO
conversions will fail on bad data.  This means that if TRY is used
with the BLANK!-in-NULL-out conventions of TO, as well as the TO
conversion itself possibly failing, that either case should cover
getting back a NULL from the length.