Switch to NULL and NONE vs. "apostrophe"
At one point, the meaning of apostrophe was a quoted NULL.

Now ~null~ is an isotopic WORD!.  ~ isotopes are used for
unset variables, and ' is a quoted void.

Whether you want to start off a field as uninitialized or
null depends on whether you want to immediately trap usages
of the variable through WORD!-reference, or if you want the
ability to test for whether the variable has been assigned
more seamlessly.

This makes some guesses, and is enough to get the tests
working.