Update LIT-WORD?! constraint to just &LIT-WORD?
With the simplification of using TYPE-WORD! for constraints, it no longer
makes sense to define another level of indirection and come up with a fake
sigil for it.  So constraints are used as type-words directly in places
like PARSE (or just functions if in type contexts already like a function
parameter spec).